URL rewrite algorithm implementation
Supports backreferences in the 'replacement' string specified in the
${n} format.

Backreference expansion is implemented using a simple state machine with
3 states:

1. COPY_INPUT - copies character by character from the 'replacement'
string to the rewritten url. If '$' is found we switch to the BACKREF_FOUND
state.

2. BACKREF_FOUND - Accepts only '{'. If found we switch to the PARSE_NUMBER
state. If not found we're going back to COPY_INPUT, and the current
character and the previous '$' are copied to the rewritten url.

3. PARSE_NUMBER - accepts only '0'..'9' and '}'. When '}' is found the
number is parsed and the corresponding replacement string is selected
from the pattern matches array. If it finds an illegal character we switch
to the COPY_INPUT state and copy all the characters ("${...") that failed to
form a valid backreference into the rewritten url.

Signed-off-by: Ionut Nicu <ionut.nicu@mindbit.ro>