Use a real domain table instead of a view
The original database design referenced the DNS zone table from the mail
domain table, but it's better to keep the two tables unrelated. The SQL
view idea and instrumenting it with a Propel "table" has some drawbacks:
  - There is no way of creating a domain without provisioning at least
    one of the DNS and mail services.
  - The implementation is not database engine agnostic.
  - There is no database relation between the DNS zone and the mail
    domain entity of the same domain.

This commit creates a real domain table that references both the DNS
zone table and the mail domain table. The domain table has a unique key
on each column that references a foreign table. This makes the relations
1:0-1.

In addition this commit adds a rudimentary mechanism for abstracting an
icon set and building image source URLs.

Signed-off-by: Radu Rendec <radu.rendec@mindbit.ro>