Rename 'domain' column to 'name' (multiple tables)
Using 'domain' as column name in the mail_domains table is not a good
idea because it makes Propel generate overlapping methods for filtering
by the 'domain' column value and by the related domains table id: both
methods are called 'filterByDomain'. Sample error:

  PHP Fatal error:  Cannot redeclare
  Mindbit\Mipanel\Model\Mipanel\Base\MailDomainQuery::filterByDomain()
  in /host/work/mipanel/model/Mipanel/Base/MailDomainQuery.php

This problem is related to the mail_domains table only, but the column
was also renamed in the domains table for consistency. Controller and
view code must be updated to match the changes (this commit breaks
existing UI functionality).

Signed-off-by: Radu Rendec <radu.rendec@mindbit.ro>